WEBAPP-208: Remove FeedbackEndpointBuilder
