--- src/modules/endpoint/MapParamsToBodyType.js
@@ -1,3 +1,0 @@
-// @flow
-
-export type MapParamsToBodyType<Params> = (params: Params) => FormData
